improve algorithm

+++ Indexing_Images_for_Content_Based_Retrieval/util.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // 返回最小数的编号 编号依次为0，1，2
-int min_index_of_three(float x, float y, float z);
+int min_index_of_three(double x, double y, double z);
